LITS-33 Use "sonar.scanner.skipJreProvisioning" in integration tests
On ephemeral CI machine this avoids unnecessary downloading and
unpacking of JRE from SQ and thus reduces time of execution of the
first project analysis in integration tests.

During execution of integration tests we already have suitable JDK.

Testing of JRE provisioning feature should not be the responsibility
of analyzers.